requirements.txt
# 中文OCR相关
paddlepaddle==2.5.0
paddleocr==2.6.1.3
opencv-python==4.8.0.76
python-poppler==0.4.1
pdf2image==1.16.3

# 数据库相关
sqlalchemy==2.0.20
pymongo==4.5.0

# 自然语言处理
jieba==0.42.1
gensim==4.3.1

# Web界面
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-wtf==1.1.1

# 工具库
numpy==1.24.3
pandas==2.0.3
tqdm==4.66.1
python-dotenv==1.0.0 